New Windows keyclient
Not linked against OpenSSL
Uses SSPI/Schanel to obtain web server cert and CA
Visual Studio .NET project, let me know if anyone wants a binary to test

Usage: keyclient

No arguments currently needed or supported. No output if it works.
Places key in the keys subdirectory of the pubcookie system dir.

TODO:  Add argument or directive to download existing key.
       More verbose error messages.

This is a major revsion and is not integrated with keyclient.c at the
moment since it only shares about 1/3 of the keyclient code.
Unclear if it would be better to #ifdef the heck out of keyclient or
leave this in a separate source file. Integration would also require
major #ifdefs to libpubcookie.c as libpubcookie.c has many links to
OpenSSL so current project duplicates the couple file I/O functions it
needs.

If we decide to leave it as a separate source, we can remove the current
Win32 #ifdef blocks from keyclient.c.


git-svn-id: https://svn.cac.washington.edu/svn/pubcookie/trunk@1153 b9756b1c-b508-4e7c-84e7-73dc1d0553b6