Add apache2 compability.
Fix some code that got clobbered in the last merge (keyserver and index.cgi)

Changed the behavior of the module install to prepend a '#' to the
LoadModule line for pubcookie.  This can be removed when the admin
adds the other pubcookie config lines.

APACHE1_3 is defined for a version 1.3 build of apache.
APACHE2 is defined for an apache 2.0 build.
APACHE is now defined in the Makefile for any module build.

I left alone a lot of places where a check for

 defined (APACHE) || defined (APACHE1_3)

can be reduced to just

 defined (APACHE)


git-svn-id: https://svn.cac.washington.edu/svn/pubcookie/trunk@1260 b9756b1c-b508-4e7c-84e7-73dc1d0553b6