Implement the new PubCookieOnDemand directive - to be documented in a separate mailing.
Implement the new PubCookieNoPrompt directive - also to be documented
in a separate mailing.  This is at present implemented in the
basic flavor only.  Will extend it to other flavors in a separate
update.

Installed a common command_rec for both apache 1.3 and 2.0 to make
maintenance easier.

Fix keyserver so that absence of keyserver_client_list means
no restriction on access.  This was the original setting and
seems to be the most flexible.


git-svn-id: https://svn.cac.washington.edu/svn/pubcookie/trunk@1272 b9756b1c-b508-4e7c-84e7-73dc1d0553b6